Normalize response metadata decisions in backend
- Extract provenance, TRACE posture, and execution event logic into a dedicated helper
- Preserve explicit retrieved chips while deriving missing values and ignore invalid TRACE axes
- Add tests for mixed chip derivation and posture normalization
